// SGUEES-SPA/src/app/pages/SelectionHiring/sc-organigrama-estructural-nivel/models/sc-organigrama-estructural-nivel.ts
export interface SC_OrganigramaEstructuralNivel {
    CORR_EMPRESA: number;
    CORR_NIVEL: number;
    NOMBRE_NIVEL: string;
    CANTIDAD_CARACTERES: number;
    ACTIVO: boolean;
    USUARIO_CREA: string;
    ESTACION_CREA: string;
    FECHA_CREA: Date;
    USUARIO_ACTU: string;
    ESTACION_ACTU: string;
    FECHA_ACTU: Date;
    EN_USO: number;
}